Updates on ms_split and its tests, adding logic
Create tests that can fully coverage the senarios in ms_split, the
possible locations for delimiter could happen in the begin, the end and
in the middle. The tests are containing 0 of those types, 1 of those
types, 2 of those types and all of those types.

The tests are fixed based on the results expected, by simply appending
the strings with a tailing '\n', the final result will append a '\n' to
the original string, and that should be the only difference. Join
function not yet implemented, not thorough tests on the number of
elements returned, what are the expected sequence of returned value.

Fix ms_split based on the bugs found from those tests

Start writing logic to parse a block of string and return mstring items
containing the records. Current idea, is to first get to the line that
following the end of a record, and begin parsing from there. The
beginning part and tailing part will be wrapped as not valid to be
merged in the merging process.